use weathux

+++ src/api/localStorage.js
@@ -1,16 +1,21 @@
-const selectedCities = 'selectedCities';
+const storageName = 'selectedCities';
 
 export const getSelectedCities = () =>
-  JSON.parse(localStorage.getItem(selectedCities)) || [];
+  JSON.parse(localStorage.getItem(storageName)) || [];
 
 export const addSelectedCity = city => {
   const cities = getSelectedCities();
-  localStorage.setItem(selectedCities, JSON.stringify([...cities, city]));
+  localStorage.setItem(
+    storageName,
+    JSON.stringify([...cities.filter(c => c.key !== city.key), city]),
+  );
 };
 
-export const initLocalStorage = () => {
+export const initLocalStorage = (force = false) => {
+  const storage = localStorage.getItem(storageName);
+  if (storage && !force) return;
   localStorage.setItem(
-    selectedCities,
+    storageName,
     JSON.stringify([
       {
         key: '2459115',
